INT-2031, INT-2112 Added support to fall back on FIFO semantics to the PriorityChannel if PRIORITY header is not provided, changed type of PRIORITY header to be Long instead of Integer

+++ spring-integration-core/src/main/java/org/springframework/integration/MessageHeaders.java
@@ -136,8 +136,8 @@ public final class MessageHeaders implements Map<String, Object>, Serializable {
 		return (sequenceSize != null ? sequenceSize : 0);
 	}
 
-	public Integer getPriority() {
-		return this.get(PRIORITY, Integer.class);
+	public Long getPriority() {
+		return this.get(PRIORITY, Long.class);
 	}
 
 	@SuppressWarnings("unchecked")
